force merged_hru values to be int type

+++ uncertainty/process.py
@@ -323,7 +323,7 @@ class UncertaintyProcess(object):
         try:
             merged_hru = geotools.read_shapefile(
                 os.path.join(self.swat_dir, 'Watershed', 'Shapes', 'hru1.shp'))
-            merged_hru = np.array(merged_hru)
+            merged_hru = np.array(merged_hru, dtype=int)
             # sort HRU_ID column (second column) into ascending order
             # and keep respective positions of other columns
             # for example:
